larger int

+++ Naped_DC_PV_SC/include/uart.h
@@ -17,7 +17,7 @@ class Uart
 
 void uart_begin(long, int);
 int uart_recive();
-void uart_recive_2_params(int *, int *);
+void uart_recive_2_params(long *, long *);
 void uart_transmit(int);
 void log_uart(String, const long *, const int);
